refact ja to ts
