+++ src/app.ts
@@ -1,17 +1,26 @@
-import { config as dotenv} from 'dotenv';
-dotenv();
 import tmi from 'tmi.js';
-import OpenCommands from './useCases/openCommands';
+import handleCommands from './useCases/handleCommands/handleCommands';
+import utils from './utils/utils';
 import config from './config';
 
-
 function handleConnected(address: any, port: any) {
-  console.log(`* Bot entrou no endereço ${address}:${port}`);
+  console.log(`*** Bot conectado com sucesso no IRC (${address}:${port}} nos canais [${options.channels}]`);
+  config.channels?.forEach(channel => {
+    //TODO refactor this
+    utils.turnOnAutomaticMessages(channel, this);
+
+    const startup = [
+      "/color yellowgreen",
+      "/me A Thasfin tá na área HeyGuys"
+    ];
+
+    startup.forEach((msg : any) => this.say(channel, msg));
+  })
 }
 
 const client = new tmi.client(config);
 
-client.on('message', OpenCommands(client));
+client.on('message', handleCommands(client));
 client.on('connected', handleConnected);
 
 client.connect();
