--- src/config.js
@@ -1,11 +1,0 @@
-module.exports = {
-    identity: {
-      username: 'thasfinbot',
-      password: process.env.TOKEN,
-    },
-    channels: ['jakeliny', 'maykbrito'],
-    startup: [
-        "/color yellowgreen",
-        "/me A Thasfin tá na área HeyGuys"
-    ],
-  };
