+++ src/useCases/initTemplates/initTemplates.ts
@@ -1,6 +1,7 @@
-const LRU = require("lru-cache");
-const path = require("path");
-const fs = require('fs').promises;
+//TODO ver o que é LRU
+import LRU from "lru-cache";
+import path from "path";
+import {promises} from 'fs';
 
 const cache = new LRU({ maxAge: 500 * 60 * 60 });
 
@@ -9,7 +10,7 @@ const cache = new LRU({ maxAge: 500 * 60 * 60 });
  * @param channel canal de onde o bot será executado
  * @returns {object|null} Retorna o template do canal ou o template padrão do bot
  */
-module.exports = async (channel = "") => {
+const initTemplates: any = async (channel = "") => {
     // Remove o primeiro valor do channel "#"
     const owner = channel.substring(1,channel.length);
 
@@ -18,10 +19,10 @@ module.exports = async (channel = "") => {
 
     // Busca os templates (default / channel) para retornar
     const [channelCommands, commonCommands] = await Promise.all([
-        fs.readFile(path.resolve(`commands.${owner}.json`), "UTF-8").catch(() => null),
-        fs.readFile(path.resolve(`commands.json`), "UTF-8").catch(() => null),
+        promises.readFile(path.resolve(`commands.${owner}.json`), "UTF-8").catch(() => null),
+        promises.readFile(path.resolve(`commands.json`), "UTF-8").catch(() => null),
     ])
-
+    //TODO ver esse buffer do channelCommands
     // Se encontrar o template do canal, prioriza o retorno e salva no cache
     if (channelCommands != null) {
         cache.set(`commands.${owner}.json`, channelCommands);
@@ -32,3 +33,5 @@ module.exports = async (channel = "") => {
     cache.set(`commands.${owner}.json`, commonCommands);
     return JSON.parse(commonCommands);
 }
+
+export default initTemplates
